table-based layout; fix non-htmlimg headlines
